# internal/app/adapter/primary/http/handler/url/add_batch.go
package urlhandler

import (
	"context"
	"encoding/json"
	"io"
	"net/http"
	"net/url"

	"go.uber.org/zap"

	dto "url-shortener/internal/app/adapter/primary/http/dto/url"
	errorshandler "url-shortener/internal/app/adapter/primary/http/handler/errors"
	"url-shortener/internal/app/adapter/primary/http/middleware"
	entity "url-shortener/internal/app/domain/entity/url"
	"url-shortener/internal/pkg/constants"
)

// AddBatch godoc
// @Summary      Пакетное создание коротких URL
// @Description  Создает несколько сокращенных URL в одном запросе.
//
// Каждый URL должен быть валидным и содержать уникальный correlation_id.
// Требуется аутентификация по API-ключу.
//
// @Security     ApiKeyAuth
// @Accept       json
// @Produce      json
// @Tags         Пакетные операции
// @Param        request body []dto.BatchRequest true "Запрос на пакетное создание URL"
//
//	example: [{"correlation_id": "1", "original_url": "https://example.com"}]
//
// @Success      201 {object} []dto.BatchResponse "Список созданных коротких URL"
//
// example: [{"correlation_id": "1", "short_url": "http://short.ly/abc123"}]
//
// @Failure      400
// @Failure      401
// @Failure      413
// @Failure      500
// @Router       /api/shorten/batch [post]
func (h *Handler) AddBatch(res http.ResponseWriter, req *http.Request) {
	ctx, cancel := context.WithTimeout(req.Context(), constants.DefaultTimeRequest)
	defer cancel()

	userID, ok := req.Context().Value(middleware.UserIDKey).(string)
	if userID == "" && !ok {
		h.logger.Error(errMessageFailedGetUserIDFromContext)
		h.handleError(res, http.StatusUnauthorized)
		return
	}

	body, err := io.ReadAll(req.Body)
	if err != nil {
		h.logger.Error(errorshandler.ErrMessageFailedReadRequestBody, zap.Error(err))
		h.handleError(res, http.StatusBadRequest)
		return
	}
	defer func() {
		_ = req.Body.Close()
	}()

	if len(body) == 0 {
		h.logger.Error(errorshandler.ErrMessageEmptyRequestBody)
		h.handleError(res, http.StatusBadRequest)
		return
	}

	var batchesRequest []dto.BatchRequest

	if err := json.Unmarshal(body, &batchesRequest); err != nil {
		h.logger.Error(errorshandler.ErrMessageFailedUnmarshalJSON, zap.Error(err))
		h.handleError(res, http.StatusBadRequest)
		return
	}

	for _, batchRequest := range batchesRequest {
		if _, err = url.ParseRequestURI(batchRequest.OriginalURL); err != nil {
			h.logger.Error(errorshandler.ErrMessageFailedParseRequestURI, zap.Error(err))
			h.handleError(res, http.StatusBadRequest)
			return
		}

		if batchRequest.ID == "" {
			h.logger.Error(errMessageEmptyBatchRequestID)
			h.handleError(res, http.StatusBadRequest)
			return
		}
	}

	//ToDo добавить методы маппинга дто, переработать ошибки, добавить метод
	var batchesEntity = make([]entity.Batch, 0, len(batchesRequest))
	for _, batchRequest := range batchesRequest {
		batch := entity.Batch{OriginalURL: batchRequest.OriginalURL, ID: batchRequest.ID}
		batchesEntity = append(batchesEntity, batch)
	}

	batches, err := h.useCase.AddBatch(ctx, batchesEntity, userID)
	if err != nil {
		h.logger.Error(errMessageFailedAddBatches, zap.Error(err))
		h.handleError(res, http.StatusInternalServerError)
		return
	}

	var batchResponse []dto.BatchResponse
	for _, batch := range batches {
		batchResponse = append(batchResponse, dto.BatchResponse{ID: batch.ID, ShortURL: batch.ShortURL})
	}

	res.Header().Set("Content-Type", "application/json")
	res.WriteHeader(http.StatusCreated)
	err = json.NewEncoder(res).Encode(batchResponse)
	if err != nil {
		h.logger.Error(errorshandler.ErrMessageFailedWriteResponse, zap.Error(err))
		res.WriteHeader(http.StatusInternalServerError)
		return
	}
}